Добавлена авторизация.

+++ Controllers/CommentController.cs
@@ -2,7 +2,9 @@
 using BlogProject.Data;
 using BlogProject.Models;
 using BlogProject.ViewModels;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace BlogProject.Controllers
 {
@@ -10,12 +12,14 @@ namespace BlogProject.Controllers
    {
       private readonly ILogger<CommentController> _logger;
       private readonly IMapper _mapper;
+      private readonly UserService _userService = new UserService();
       private readonly CommentService _commentService = new CommentService();
       public CommentController(ILogger<CommentController> logger, IMapper mapper)
       {
          _logger = logger;
          _mapper = mapper;
       }
+      [Authorize(Roles = "moder")]
       [HttpGet]
       [Route("GetComments")]
       public async Task<Comment[]> GetComments()
@@ -23,6 +27,7 @@ namespace BlogProject.Controllers
          var comments = await _commentService.GetComments();
          return comments;
       }
+      [Authorize(Roles = "user")]
       [HttpGet]
       [Route("GetCommentById")]
       public async Task<Comment> GetCommentById(Guid id)
@@ -30,36 +35,64 @@ namespace BlogProject.Controllers
          var comment = await _commentService.GetCommentById(id);
          return comment;
       }
+      [Authorize(Roles = "user")]
       [HttpPost]
       [Route("AddComment")]
-      public async Task<IActionResult> AddComment(Guid UserId, Guid PostId, AddCommentViewModel view)
+      public async Task<IActionResult> AddComment(Guid postId, AddCommentViewModel view)
       {
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userId = user.Id;
+
          var comment = _mapper.Map<AddCommentViewModel, Comment>(view);
-         comment.UserId = UserId;
-         comment.PostId = PostId;
+         comment.UserId = userId;
+         comment.PostId = postId;
          await _commentService.Save(comment);
-         return StatusCode(200, "Успех");
+         return Ok();
       }
+      [Authorize(Roles = "user")]
       [HttpPut]
       [Route("EditComment")]
       public async Task<IActionResult> EditComment(Guid id, AddCommentViewModel view)
       {
-         var comment = await GetCommentById(id);
-         if (comment == null)
-            return StatusCode(200, "Не Успех");
-         var newComment = _mapper.Map<AddCommentViewModel, Comment>(view);
-         await _commentService.Update(comment, newComment);
-         return StatusCode(200, "Успех");
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var claimRoles = ident.Claims.Where(u => u.Type == ClaimTypes.Role).ToArray();
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userComments = await _commentService.GetCommentByUserId(user.Id);
+
+         if (userComments.FirstOrDefault(p => p.Id == id) != null || claimRoles.FirstOrDefault(r => r.Value == "moder") != null)
+         {
+            var comment = await GetCommentById(id);
+            if (comment == null)
+               return BadRequest();
+            var newComment = _mapper.Map<AddCommentViewModel, Comment>(view);
+            await _commentService.Update(comment, newComment);
+            return Ok();
+         }
+         return BadRequest();
       }
+      [Authorize(Roles = "user")]
       [HttpDelete]
       [Route("DeleteComment")]
       public async Task<IActionResult> DeleteComment(Guid id)
       {
-         var comment = await GetCommentById(id);
-         if (comment == null)
-            return StatusCode(200, "Не Успех");
-         await _commentService.Delete(comment);
-         return View();
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var claimRoles = ident.Claims.Where(u => u.Type == ClaimTypes.Role).ToArray();
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userComments = await _commentService.GetCommentByUserId(user.Id);
+
+         if (userComments.FirstOrDefault(p => p.Id == id) != null || claimRoles.FirstOrDefault(r => r.Value == "moder") != null)
+         {
+            var comment = await GetCommentById(id);
+            if (comment == null)
+               return BadRequest();
+            await _commentService.Delete(comment);
+            return Ok();
+         }
+         return BadRequest();
       }
    }
 }
