+++ Controllers/PostController.cs
@@ -2,8 +2,11 @@
 using BlogProject.Data;
 using BlogProject.Models;
 using BlogProject.ViewModels;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Data;
+using System.Security.Claims;
 
 namespace BlogProject.Controllers
 {
@@ -11,6 +14,7 @@ namespace BlogProject.Controllers
    {
       private readonly ILogger<PostController> _logger;
       private readonly IMapper _mapper;
+      private readonly UserService _userService = new UserService();
       private readonly PostService _postService = new PostService();
       private readonly TagService _tagService = new TagService();
       private readonly CommentService _commentService = new CommentService();
@@ -19,6 +23,7 @@ namespace BlogProject.Controllers
          _logger = logger;
          _mapper = mapper;
       }
+      [Authorize(Roles = "user")]
       [HttpGet]
       [Route("GetPosts")]
       public async Task<PostViewModel[]> GetPosts()
@@ -35,6 +40,7 @@ namespace BlogProject.Controllers
          }
          return postViewModels;
       }
+      [Authorize(Roles = "user")]
       [HttpGet]
       [Route("GetPostById")]
       public async Task<PostViewModel> GetPostById(Guid id)
@@ -47,6 +53,7 @@ namespace BlogProject.Controllers
          postViewModel.Comments = await _commentService.GetCommentByPostId(post.Id);
          return postViewModel;
       }
+      [Authorize(Roles = "user")]
       [HttpGet]
       [Route("GetPostByUserId")]
       public async Task<PostViewModel[]> GetPostByUserId(Guid id)
@@ -63,35 +70,63 @@ namespace BlogProject.Controllers
          }
          return postViewModels;
       }
+      [Authorize(Roles = "user")]
       [HttpPost]
       [Route("AddPost")]
-      public async Task<IActionResult> AddPost(Guid userId, Guid[] tagIds, AddPostViewModel view)
+      public async Task<IActionResult> AddPost(Guid[] tagIds, AddPostViewModel view)
       {
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userId = user.Id;
+
          var post = _mapper.Map<AddPostViewModel, Post>(view);
          post.UserId = userId;
          await _postService.Save(post, tagIds);
-         return StatusCode(200, "Успех");
+         return Ok();
       }
+      [Authorize(Roles = "user")]
       [HttpPut]
       [Route("EditPost")]
       public async Task<IActionResult> EditPost(Guid id, Guid[] tagIds, AddPostViewModel view)
       {
-         var editPost = await _postService.GetPostById(id);
-         if (editPost == null)
-            return StatusCode(200, "Не Успех");
-         var newPost = _mapper.Map<AddPostViewModel, Post>(view);
-         await _postService.Update(editPost, newPost, tagIds);
-         return StatusCode(200, "Успех");
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var claimRoles = ident.Claims.Where(u => u.Type == ClaimTypes.Role).ToArray();
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userPosts = await _postService.GetPostByUserId(user.Id);
+
+         if (userPosts.FirstOrDefault(p => p.Id == id) != null || claimRoles.FirstOrDefault(r => r.Value == "moder") != null)
+         {
+            var editPost = await _postService.GetPostById(id);
+            if (editPost == null)
+               return BadRequest();
+            var newPost = _mapper.Map<AddPostViewModel, Post>(view);
+            await _postService.Update(editPost, newPost, tagIds);
+            return Ok();
+         }
+         return BadRequest();
       }
+      [Authorize(Roles = "user")]
       [HttpDelete]
       [Route("DeletePost")]
       public async Task<IActionResult> DeletePost(Guid id)
       {
-         var post = await _postService.GetPostById(id);
-         if (post == null)
-            return StatusCode(200, "Не Успех");
-         await _postService.Delete(post);
-         return StatusCode(200, "Успех");
+         ClaimsIdentity ident = HttpContext.User.Identity as ClaimsIdentity;
+         var claimEmail = ident.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Name).Value;
+         var claimRoles = ident.Claims.Where(u => u.Type == ClaimTypes.Role).ToArray();
+         var user = await _userService.GetUserByEmail(claimEmail);
+         var userPosts = await _postService.GetPostByUserId(user.Id);
+
+         if (userPosts.FirstOrDefault(p => p.Id == id) != null || claimRoles.FirstOrDefault(r => r.Value == "moder") != null)
+         {
+            var post = await _postService.GetPostById(id);
+            if (post == null)
+               return BadRequest();
+            await _postService.Delete(post);
+            return Ok();
+         }
+         return BadRequest();
       }
    }
 }
